Move dns_utils to pkg/shared
